requirements.txt
azure-cli-core==2.19.1
azure-mgmt-resource==15.0.0
azure-mgmt-storage==16.0.0
black==20.8b1
flake8==3.8.4
pydocstyle==5.1.1
pylint==2.6.0
pytest-cov==2.11.1
wheel==0.36.2
